New Activities for 3 Modules; Standard format for 3 more
Minor corrections for To the Front; still not ready to publish (placeholder for one activity).
New activities for I Could Do That; still not ready to publish.
Minor changes to Locomotive; still not ready to publish (placeholder for one activity).
Standardized format for three modules: Freedom on the Menu; Grace Banker; The Gardener. These are ready to publish (no new content, just moved existing activities to the standard, fixed format for Supp. Reading activity)